Feature: Have `cosign_sign` warn when ambient credentials are missing. (#6)
:gift: This makes the absence of ambient credentials non-fatal at deploy
time, which is a nice balance for use cases that don't want signing when
deploying from a workstation, but do from release automation (e.g. a
shared module btw dev envs and producation). This can be made fatal by
threading the image through `cosign_verify`, and in the future we might
make this a knob (with diff defaults), but this feels like a reasonable
balance to start.

/kind feature