Push rate limiting down into rekor client
This requires us to fork a small package from sigstore/rekor, but it
gives us a few benefits:

We rate limit per HTTP request, which is more aligned with the actual
rate limit (rather than per rekor client method call).

We rate limit the innermost transport, which means retries won't bypass
the limit.

We can modify the default transport to use a pooled transport, which
should reuse connections, which should be faster with concurrent
requests.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>