Pass a nop verifier to dupe detection (#81)
I believe the dupe detector takes a verifier so that it only replaces
attestations that match a verifier. The verifier we pass into Attest is
ephemeral and probably not what the dupe detector really wants. For
attestations that fail to verify, the dupe detector just ignores the
error and continues.

For our use case, we always want to replace by predicate type and don't
care about keeping around multiple identities/verifiers. We _want_ to
clobber whatever was there before and update it with the most recent
identity.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>