Fork cosign's replace op stuff (#87)
The way this is written in cosign, we re-fetch and parse each layer for
every call to Replace. Since we're replacing 3 things with 3 new things,
we end up having to fetch and parse 9 blobs, which can take several
seconds altogether. Fixing this in cosign could be done by memoizing the
predicateType, but we have a further narrowing case where we add the
predicateType as an annotation on each attestation, so we can do a very
fast lookup of that and skip fetching the blob altogether.

This also drops the dupe checker because it's redundant with the replace
operation and incurs a little bit of overhead we would rather avoid.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>