Do not re-attest things that have not changed
This does some diffing of the existing statements with the new desired
statements and omits any that do not need to be re-attested. This should
reduce the amount of traffic we need to send to rekor and thus the
amount of self-throttling we do, which is currently a huge bottleneck on
moduels with a large number of cosign_attests.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>