Add conflict to cosign_sign
The default is APPEND, which matches the current behavior today. This
adds both REPLACE and SKIPSAME to match what cosign_attest does. Instead
of using predicateType as the replacement criteria, this uses the
signature layer's digest.

It should be safe to bump this before we bump terraform-publisher-apko
because the default matches the current behavior.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>