Extract SignEntity and AttestEntity functions (#316)
This can be used by callers to add signatures and attestations to an
image without interacting with the registry. This is useful if there are
several related operations that need to be performed on an
`oci.SignedEntity` without round-tripping to the registry.

This doesn't change any of the existing logic. It just moves things
around to expose the new functions.