Make SKIPSAME skip rekor upload when signature isn't needed (#320)
Remove unecessary calls to Rekor. When signing an image with SKIPSAME,
we currently:
* *First* produce an `oci.Signature` using the `rekor.Cosigner`. This
has the side-effect of uploading an entry into the Rekor transparency
log.
* *Then* look for duplicate signatures on the image to decide if the new
signature should be attached.

This means even if a duplicate is found, we've already unnecessarily
uploaded information to Rekor. This PR re-orders logic so that we first
check for a duplicate and only proceed further if one is not found.