Change Rekor rate limiting to be context-aware (#334)
There are two aspects to this change:

1) We're now using a rate limiter that is aware of the context deadline
+ context cancellation

The prior implementation effectively just did a `Sleep()` to maintain
the requested rate limit. This causes problems if the wait time falls
outside of the context lifetime. The goroutine will wait, but it's not
going to be able to do anything useful when it wakes up. This can easily
cascade into a situation where all goroutines are waiting even after
their context has been cancelled.

The rate limiter from `x/time/rate` is aware of the context and will
immediately return an error rather than waiting past its deadline. It'll
also cancel its "reservation" in the rate limiter if the context is
cancelled for other reasons, possibly letting another goroutine proceed
with a shorter wait.

2) Move the rate limiter check into `SignEntity()/AttestEntity()`

The prior implementation applied rate limiting on each individual HTTP
request going to Rekor. Now that the call to the rate limiter may fail,
this means attesting 5 statements may succeed for the first 4 but fail
for the last one. This wastes our calls to Rekor, since we won't
actually end up using the result. We can avoid this type of partial
failure by waiting up-front for the number of statements we need to
sign/upload.