Call rekor rate limiter before signing payload
The current implementation of SignEntity signs the payload, throttles with a rate limiter, then finally calls Rekor. Depending on how long the rate limiter waits, it's possible the signature will be invalid when Rekor is called. If this happens, the image will fail validation with `cosign`. It's better to instead throttle ourselves upfront and then quickly sign+upload to minimize the chance of the signature expiring.